Moved more tests for transformations
